Split code for go-1.8 and go-1.7
